+++ lib/src/taro_loader.dart
@@ -30,16 +30,13 @@ class TaroLoader {
 
   /// Loads the data from the provided URL with the given request headers.
   /// If [checkMaxAgeIfExist] is true, the method checks the max age of the data.
-  /// The [resizeOption] parameter is used to resize the image. If it is not provided, the default resize option is used.
   Future<Uint8List> load({
     required String url,
     required Map<String, String> headers,
-    required TaroResizeOption resizeOption,
     required TaroHeaderOption headerOption,
   }) async {
     final storageBytes = await _storageLoader.load(
       url: url,
-      resizeOption: resizeOption,
     );
     if (storageBytes != null) {
       return storageBytes;
@@ -48,7 +45,6 @@ class TaroLoader {
     final networkResponse = await _networkLoader.load(
       url: url,
       headers: headers,
-      resizeOption: resizeOption,
       headerOption: headerOption,
     );
 
@@ -64,7 +60,6 @@ class TaroLoader {
       bytes: networkResponse.bytes,
       contentType: networkResponse.contentType,
       expireAt: networkResponse.expireAt,
-      resizeOption: resizeOption,
     );
 
     return networkResponse.bytes;